refactor: rename folder shared to common
relates to #2
