+++ src/index.tsx
@@ -5,9 +5,9 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import App from './app/App';
+import { WaitingProvider } from './common/wait/WaitContext';
 import configureStore, { history } from './configureStore';
 import * as serviceWorker from './serviceWorker';
-import { WaitingProvider } from './shared/wait/WaitContext';
 
 // Add ES6 Map support for redux-devtools-extension
 // See: https://github.com/zalmoxisus/redux-devtools-extension/issues/124
